+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 ## 0.5.0
 
+- Add 'TimeStamp' type class to abstract working with types that have an
+  absolute time associated with them.
 - Add 'TimeInterval' type for working with intervals of time, there are also
   some helper functions to make it easier to work with intervals.
 - Extend `ModelParameters` class to have an `eventWeights` to provide a vector